Move get_args to functional
