+++ corebehrt/main/finetune_cv.py
@@ -7,10 +7,12 @@ import torch
 from corebehrt.modules.setup.config import load_config
 from corebehrt.modules.setup.manager import ModelManager
 from corebehrt.modules.setup.loader import load_and_select_splits
-from corebehrt.modules.setup.directory import DirectoryPreparer, get_args
+from corebehrt.modules.setup.directory import DirectoryPreparer
+from corebehrt.functional.setup.args import get_args
 from corebehrt.modules.monitoring.metric_aggregation import (
     compute_and_save_scores_mean_std,
 )
+
 from corebehrt.functional.load import load_pids
 
 from corebehrt.functional.features.split import (
